+++ Sources/ConControlsTests/UnitTests/Controls/KeyCombination/Without.cs
@@ -1,0 +1,22 @@
+﻿/*
+ * (C) René Vogt
+ *
+ * Published under MIT license as described in the LICENSE.md file.
+ *
+ */
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+#nullable enable
+
+namespace ConControlsTests.UnitTests.Controls.KeyCombination
+{
+    public partial class KeyCombinationTests
+    {
+        [TestMethod]
+        public void Without_Inconclusive()
+        {
+            Assert.Inconclusive();
+        }
+    }
+}