+++ Sources/ConControlsTests/UnitTests/Helpers/KeyHandlingExtensions/Matches.cs
@@ -1,0 +1,21 @@
+﻿/*
+ * (C) René Vogt
+ *
+ * Published under MIT license as described in the LICENSE.md file.
+ *
+ */
+
+#nullable enable
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace ConControlsTests.UnitTests.Helpers.KeyHandlingExtensions
+{
+    public partial class KeyHandlingExtensionsTests
+    {
+        [TestMethod]
+        public void Matches_Inconcusive()
+        {
+        }
+    }
+}